many changes but simple. These are mostly changes to ensure that the demos and tests still compile and run after several changes that were made to remove warnings.

+++ src/GEL/CGLA/eigensolution.cpp
@@ -40,7 +40,7 @@ namespace CGLA
         {
             // Seed the eigenvector estimate
             VT q;
-            for (size_t j=0; j<MT::get_v_dim(); ++j) 
+            for (unsigned int j=0; j<MT::get_v_dim(); ++j)
                 q[j] = gel_rand()/static_cast<double>(GEL_RAND_MAX);
             
             q.normalize();
